refactor(macros): split DIContainerCodeGenerator into scope-based files (N-1)
Splits the 1,259-line `DIContainerCodeGenerator.swift` into four
focused files, keeping the public entry points and common helpers in
the main file and moving each codegen concern to its own unit. The
split is purely mechanical — macro expansion snapshots stay
byte-identical and all 349 existing tests pass without record
refreshes.

New files:

- `DIContainerOverridesGenerator.swift` — `overrideCandidateMembers`,
  `makeOverridesStructDecl`, `makeConvenienceInitDecl`. Owns the
  nested `Overrides` struct and its trailing-closure convenience init.
  Phase M's two extra slots per `@SubContainer` member ride along
  unchanged.
- `DIContainerWithOverridesGenerator.swift` —
  `makeWithOverridesMethods` / `makeWithOverridesMethod`. The four
  `sync` / `throws` / `async` / `async throws` effect overloads live
  together so future Sendable / concurrency tightening (Phase N-2)
  can be reviewed in isolation.
- `DIContainerSubContainerGenerator.swift` —
  `makeSubContainerInitStatements`,
  `subContainerSharedAssignmentExpr`, `subContainerInitializerExpr`,
  `resolvedSubContainerArguments`, plus the Phase M `if let direct`
  ladder helper. Keeps the `.shared` / `.transient` branching and
  `_innoDISubBuild_` closure capture close to the `@SubContainer`
  expansion model.

Shared helpers promoted to `internal` so the new files can call in
without additional plumbing:

- `accessModifiers(_:)`, `mainActorAttributeList()`,
  `optionalParameterType(for:)`, `assignExpr(targetName:valueName:)`,
  `assignExprWithValue(targetName:value:)`.

The remaining `DIContainerCodeGenerator.swift` (656 lines) still
hosts the primary `generateInit` / `generateAll` entry points and
the init-body helpers (`makeInitDecl`, `makeFactoryExpr`,
`makeAsyncFactoryExpr`, `closureArgumentExpressions`,
`dependencyExpression`, `resolveClosureParameter`, …) that exclusively
belong to the primary initializer construction. No behavior changes.
